requirements.txt
google-cloud-storage
transformers
datasets
huggingsound
pandas
soundfile
librosa
ffmpeg-python
pygobject
evaluate